requirements.txt
python-docx==0.8.11
mysql-connector-python==8.0.33